+++ NEWS.md
@@ -1,4 +1,4 @@
-# semhelpinghands 0.1.3.0
+# semhelpinghands 0.1.4.0
 
 - Improve the notes in examples. (0.1.2.1)
 - Fix some bugs. (0.1.2.1)
@@ -7,6 +7,8 @@
 - Add `se_ratios()` for comparing standard errors by
   different estimators. (0.1.2.4)
 - Add `show_more_options()`. (0.1.2.5)
+- Fix a bug in estimates table manipulation functions. (0.1.3.2)
+- Fix a bug in the test that used known results. (0.1.4.0)
 
 # semhelpinghands 0.1.2.0